+++ data-raw/ucdp_acd.R
@@ -28,6 +28,21 @@ ucdp_acd %>%
   mutate_at(vars("start_date", "start_date2", "ep_end_date"), ~lubridate::ymd(.)) %>%
   mutate_at(vars("gwno_a", "gwno_a_2nd", "gwno_b", "gwno_b_2nd"), ~as.numeric(.)) -> ucdp_acd
 
+ucdp_acd %>%
+  mutate(type_of_conflict = case_when(
+    type_of_conflict == 1 ~ "extrasystemic",
+    type_of_conflict == 2 ~ "interstate",
+    type_of_conflict == 3 ~ "intrastate",
+    type_of_conflict == 4 ~ "II",
+  )) -> ucdp_acd
+
+ucdp_acd %>%
+  mutate(incompatibility = case_when(
+    incompatibility == 1 ~ "territory",
+    incompatibility == 2 ~ "government",
+    incompatibility == 3 ~ "both"
+  )) -> ucdp_acd
+
 save(ucdp_acd, file="data/ucdp_acd.rda")
 
 
@@ -43,13 +58,61 @@ ucdp_acd %>%
                            gwcode == gwno_a_2nd ~ 1,
                            TRUE ~ 0)) %>%
   select(conflict_id, year, gwcode, sidea, everything()) %>%
-  arrange(year, conflict_id)
+  arrange(year, conflict_id) %>%
   filter(type_of_conflict == 2) %>% summary
+
+
+ucdp_acd %>%
+  distinct(start_date2)
+
+
+
+ucdp_acd %>%
+  filter(is.na(gwno_b)) %>%
+  distinct(type_of_conflict)
+
 
 ucdp_acd %>%
   filter(type_of_conflict %in% c(1, 3)) %>% summary
 
 
 ucdp_acd %>%
+  filter(!is.na(gwno_b))
+
+
+ucdp_acd %>%
   filter(conflict_id == 11345) %>%
   data.frame
+
+
+
+ucdp_acd %>%
+  filter(type_of_conflict == 3) %>%
+  mutate(epstyear = lubridate::year(start_date2),
+         ependyear = lubridate::year(ep_end_date)) %>%
+  mutate()
+
+
+
+
+
+create_stateyears(system = "gw") %>%
+  add_ucdp_acd()
+  group_by(conflict_id, gwno_a, year) %>% slice(1) %>%
+  arrange(conflict_id, gwno_a, year) %>%
+  group_by(conflict_id, gwno_a) %>%
+  mutate(ucdponset = ifelse(row_number() == 1, 1, 0)) %>%
+  group_by(gwno_a, year) %>%
+  summarize(ucdpongoing = 1,
+            maxintensity = max(intensity_level),
+            ucdponset = max(ucdponset),
+            conflict_ids = paste0(conflict_id, collapse = "; ")) %>%
+  ungroup() %>%
+  rename(gwcode = gwno_a) %>%
+  left_join(create_stateyears() %>% filter(year >= 1945), .)
+
+ucdp_acd %>%
+  select(gwno_a,  start_date2, ep_end_date)
+
+
+