+++ NEWS.md
@@ -2,6 +2,7 @@ peacesciencer 0.5.0 (in development)
 ---------------------------------------------------------------------
 
 - Add `add_strategic_rivalries()`.
+- Add `add_ucdp_acd()`. With it, update `ucdp_acd` data for functionality.
 - Clarify documentation for `add_sdp_gdp()` about constructing a GDP per capita estimate.
 - Clarify documentation for `add_contiguity()` for why non-contiguous dyads are 0s and not 6s. Briefly: I don't think of these data as ordinal and I don't encourage the researcher to think of them as ordinal either.
 - Fix bug in `add_nmc()` (and `cow_nmc`). -9 is now explicitly NA, as it should have been.